refactor: dry graphql lint custom rules
